+++ imio/dms/mail/content/behaviors.py
@@ -279,3 +279,14 @@ class PlonegroupUserLinkUseridValidator(validator.SimpleFieldValidator):
 
 
 validator.WidgetValidatorDiscriminators(PlonegroupUserLinkUseridValidator, field=IPlonegroupUserLink['userid'])
+
+
+@provider(IFormFieldProvider)
+class IDefaultContentCategoryBehavior(model.Schema):
+
+    default_content_category = schema.Choice(
+        title=_(u"Default content category"),
+        description=_("Category to apply on generated documents from this template."),
+        vocabulary=u"imio.dms.mail.PODTemplateContentCategoriesVocabulary",
+        required=True,
+    )